refactor(fc-driver): the quiesced flag tracks the guest through a checkpoint; handle tests in their own file
After a successful pause the handle reports Quiesced even if a later
resume fails, instead of Running while the guest is frozen.
